add test for min/max int, change return types to long int, add abs value for min check

+++ Module08/ex01/includes/Span.hpp
@@ -21,11 +21,10 @@ class	Span
 		Span&	operator=( const Span& to_copy );
 
 		/* PUBLIC METHODS */
-		void	addNumber( int to_add ) throw(std::length_error);
-		int		shortestSpan( void ) throw(std::length_error);
-		int		longestSpan( void ) throw(std::length_error);
-		void	addManyNumbers(int to_add[], size_t size) throw(std::length_error);
-
+		void		addNumber( int to_add ) throw(std::length_error);
+		long int		shortestSpan( void ) throw(std::length_error);
+		long int		longestSpan( void ) throw(std::length_error);
+		void		addManyNumbers(int to_add[], size_t size) throw(std::length_error);
 };
 
 #endif
